move fixtures to new conftest
